prepared to improve qt config
will add features to access and write/read/overwrite configurations in the qteasy.cfg file
