+++ qteasy/core.py
@@ -24,8 +24,7 @@ from .finance import CashPlan
 from .qt_operator import Operator
 from .visual import _plot_loop_result, _loop_report_str, _print_test_result
 from .visual import _plot_test_result
-from ._arg_validators import _update_config_kwargs, ConfigDict
-from ._arg_validators import QT_CONFIG, _vkwargs_to_text
+from ._arg_validators import QT_CONFIG, ConfigDict
 from .optimization import _evaluate_all_parameters, _evaluate_one_parameter
 
 
@@ -1096,479 +1095,6 @@ def help(**kwargs):
     -------
     """
     raise NotImplementedError
-
-
-def configure(config=None, reset=False, only_built_in_keys=True, **kwargs) -> None:
-    """ 配置qteasy的运行参数QT_CONFIG
-
-    Parameters
-    ----------
-    config: ConfigDict 对象
-        需要设置或调整参数的config对象，默认为None，此时直接对QT_CONFIG对象设置参数
-    reset: bool
-        默认值为False，为True时忽略传入的kwargs，将所有的参数设置为默认值
-    only_built_in_keys: bool
-        默认值False，如果为True，仅允许传入内部参数，False时允许传入任意参数
-    **kwargs:
-        需要设置的所有参数
-
-    Returns
-    -------
-    None
-
-    Notes
-    -----
-    使用get_config()或configuration()查看当前的QT_CONFIG参数
-
-    Examples
-    --------
-    >>> configure(reset=True)  # 将QT_CONFIG参数设置为默认值
-    >>> configure(invest_cash_amounts=[10000, 20000, 30000], invest_cash_dates=['2018-01-01', '2018-02-01', '2018-03-01'])
-    >>> get_config('invest_cash_amounts, invest_cash_dates')
-    No. Config-Key            Cur Val                       Default val
-    -------------------------------------------------------------------
-    1   invest_cash_amounts   [10000, 20000, 30000]         <[100000.0]>
-    2   invest_cash_dates     ['2018-01-01', '2018-02-01'...<None>
-
-    """
-    if config is None:
-        set_cfg = QT_CONFIG
-    else:
-        assert isinstance(config, ConfigDict), TypeError(f'config should be a ConfigDict, got {type(config)}')
-        set_cfg = config
-    if not reset:
-        _update_config_kwargs(set_cfg, kwargs, raise_if_key_not_existed=only_built_in_keys)
-    else:
-        from qteasy._arg_validators import _valid_qt_kwargs
-        default_kwargs = {k: v['Default'] for k, v in zip(_valid_qt_kwargs().keys(),
-                                                          _valid_qt_kwargs().values())}
-        _update_config_kwargs(set_cfg, default_kwargs, raise_if_key_not_existed=True)
-
-
-def set_config(config=None, reset=False, only_built_in_keys=True, **kwargs) -> None:
-    """ 配置qteasy的运行参数QT_CONFIG，等同于configure()
-
-    Parameters
-    ----------
-    config: ConfigDict 对象
-        需要设置或调整参数的config对象，默认为None，此时直接对QT_CONFIG对象设置参数
-    reset: bool
-        默认值为False，为True时忽略传入的kwargs，将所有的参数设置为默认值
-    only_built_in_keys: bool
-        默认值False，如果为True，仅允许传入内部参数，False时允许传入任意参数
-    **kwargs:
-        需要设置的所有参数
-
-    Returns
-    -------
-    None
-
-    Examples
-    --------
-    参见 configure()
-    """
-
-    return configure(config=config, reset=reset, only_built_in_keys=only_built_in_keys, **kwargs)
-
-
-def configuration(config_key=None, level=0, up_to=0, default=True, verbose=False) -> None:
-    """ 显示qt当前的配置变量，
-
-    Parameters
-    ----------
-    config_key : str or list of str
-        需要显示的配置变量名称，如果不给出任何名称，则按level，up_to等方式显示所有的匹配的变量名
-        可以以逗号分隔字符串的形式给出一个或多个变量名，也可以list形式给出一个或多个变量名
-        以下两种方式等价：
-        'local_data_source, local_data_file_type, local_data_file_path'
-        ['local_data_source', 'local_data_file_type', 'local_data_file_path']
-    level : int, Default: 0
-        需要显示的配置变量的级别。
-        如果给出了config，则忽略此参数
-    up_to : int, Default: 0
-        需要显示的配置变量的级别上限，需要配合level设置。
-        例如，当level == 0, up_to == 2时
-        会显示级别在0～2之间的所有配置变量
-        如果给出了config，则忽略此参数
-    default: Bool, Default: False
-        是否显示配置变量的默认值，如果True，会同时显示配置变量的当前值和默认值
-    verbose: Bool, Default: False
-        是否显示完整说明信息，如果True，会同时显示配置变量的详细说明
-
-    Returns
-    -------
-    None
-
-    Notes
-    -----
-    使用示例参见get_config()
-    """
-    assert isinstance(level, int) and (0 <= level <= 5), f'InputError, level should be an integer, got {type(level)}'
-    assert isinstance(up_to, int) and (0 <= up_to <= 5), f'InputError, up_to level should be an integer, ' \
-                                                         f'got {type(up_to)}'
-    if up_to <= level:
-        up_to = level
-    if up_to == level:
-        level = level
-    else:
-        level = list(range(level, up_to + 1))
-
-    if config_key is None:
-        kwargs = QT_CONFIG
-    else:
-        if isinstance(config_key, str):
-            config_key = str_to_list(config_key)
-        if not isinstance(config_key, list):
-            raise TypeError(f'config_key should be a string or list of strings, got {type(config_key)}')
-        assert all(isinstance(item, str) for item in config_key)
-        kwargs = {key: QT_CONFIG[key] for key in config_key}
-        level = [0, 1, 2, 3, 4, 5]
-
-    from shutil import get_terminal_size
-    screen_width = get_terminal_size().columns
-    from rich import print as rprint
-    rprint(_vkwargs_to_text(
-            kwargs=kwargs,
-            level=level,
-            info=default,
-            verbose=verbose,
-            width=screen_width,
-    ))
-
-
-def get_configurations(config_key=None, level=0, up_to=0, default=True, verbose=False):
-    """ 显示qt当前的配置变量，与get_config / configuration 相同
-
-    Parameters
-    ----------
-    config_key : str or list of str
-        需要显示的配置变量名称，如果不给出任何名称，则按level，up_to等方式显示所有的匹配的变量名
-        可以以逗号分隔字符串的形式给出一个或多个变量名，也可以list形式给出一个或多个变量名
-        以下两种方式等价：
-        'local_data_source, local_data_file_type, local_data_file_path'
-        ['local_data_source', 'local_data_file_type', 'local_data_file_path']
-    level : int, Default: 0
-        需要显示的配置变量的级别。
-        如果给出了config，则忽略此参数
-    up_to : int, Default: 0
-        需要显示的配置变量的级别上限，需要配合level设置。
-        例如，当level == 0, up_to == 2时
-        会显示级别在0～2之间的所有配置变量
-        如果给出了config，则忽略此参数
-    default: Bool, Default: False
-        是否显示配置变量的默认值，如果True，会同时显示配置变量的当前值和默认值
-    verbose: Bool, Default: False
-        是否显示完整说明信息，如果True，会同时显示配置变量的详细说明
-
-    Returns
-    -------
-    None
-
-    Examples
-    --------
-    使用示例参见get_config()
-    """
-
-    return configuration(config_key=config_key, level=level, up_to=up_to, default=default, verbose=verbose)
-
-
-def get_config(config_key=None, level=0, up_to=0, default=True, verbose=False):
-    """ 显示qt当前的配置变量，与get_config / configuration 相同
-
-    Parameters
-    ----------
-    config_key : str or list of str
-        需要显示的配置变量名称，如果不给出任何名称，则按level，up_to等方式显示所有的匹配的变量名
-        可以以逗号分隔字符串的形式给出一个或多个变量名，也可以list形式给出一个或多个变量名
-        以下两种方式等价：
-        'local_data_source, local_data_file_type, local_data_file_path'
-        ['local_data_source', 'local_data_file_type', 'local_data_file_path']
-    level : int, Default: 0
-        需要显示的配置变量的级别。
-        如果给出了config，则忽略此参数
-    up_to : int, Default: 0
-        需要显示的配置变量的级别上限，需要配合level设置。
-        例如，当level == 0, up_to == 2时
-        会显示级别在0～2之间的所有配置变量
-        如果给出了config，则忽略此参数
-    default: Bool, Default: False
-        是否显示配置变量的默认值，如果True，会同时显示配置变量的当前值和默认值
-    verbose: Bool, Default: False
-        是否显示完整说明信息，如果True，会同时显示配置变量的详细说明
-
-    Returns
-    -------
-    None
-
-    Examples
-    --------
-    >>> get_config('local_data_source')
-    No. Config-Key            Cur Val        Default val
-    ----------------------------------------------------
-    1   local_data_source     database       <file>
-
-    >>> get_config('local_data_source, local_data_file_type, local_data_file_path')
-    No. Config-Key            Cur Val        Default val
-    ----------------------------------------------------
-    1   local_data_source     database       <file>
-    2   local_data_file_type  csv            <csv>
-    3   local_data_file_path  data/          <data/>
-
-    >>> get_config(level=0, up_to=2)
-    No. Config-Key            Cur Val        Default val
-    ----------------------------------------------------
-    1   mode                  1              <1>
-    2   time_zone             Asia/Shanghai  <Asia/Shanghai>
-    3   asset_pool            000300.SH      <000300.SH>
-    4   asset_type            IDX            <IDX>
-    5   live_trade_account_id None           <None>
-    6   live_trade_account    None           <None>
-    7   live_trade_debug_mode False          <False>
-    8   live_trade_init_cash  1000000.0      <1000000.0>
-    ... (more rows)
-
-    >>> get_config(level=0, up_to=1, verbose=True)
-    No. Config-Key            Cur Val        Default val
-          Description
-    ----------------------------------------------------
-    1   mode                  1              <1>
-          qteasy 的运行模式:
-          0: 实盘运行模式
-          1: 回测-评价模式
-          2: 策略优化模式
-          3: 统计预测模式
-    2   time_zone             Asia/Shanghai  <Asia/Shanghai>
-          回测时的时区，可以是任意时区，例如：
-          Asia/Shanghai
-          Asia/Hong_Kong
-          US/Eastern
-          US/Pacific
-          Europe/London
-          Europe/Paris
-          Australia/Sydney
-          Australia/Melbourne
-          Pacific/Auckland
-          Pacific/Chatham
-          etc.
-    3   asset_pool            000300.SH      <000300.SH>
-          可用投资产品池，投资组合基于池中的产品创建
-    4   asset_type            IDX            <IDX>
-          投资产品的资产类型，包括：
-          IDX  : 指数
-          E    : 股票
-          FT   : 期货
-          FD   : 基金
-    """
-
-    return configuration(config_key=config_key, level=level, up_to=up_to, default=default, verbose=verbose)
-
-
-def _check_config_file_name(file_name, allow_default_name=False):
-    """ 检查配置文件名是否合法，如果合法或可以转化为合法，返回合法文件名，否则raise
-
-    parameters
-    ----------
-    file_name: str
-        配置文件名，可以是一个文件名，也可以是不带后缀的文件名，不可以是一个路径
-    allow_default_name: bool
-        是否允许使用默认的配置文件名qteasy.cfg
-    """
-
-    if file_name is None:
-        file_name = 'saved_config.cfg'
-    if not isinstance(file_name, str):
-        raise TypeError(f'file_name should be a string, got {type(file_name)} instead.')
-    import re
-    if re.match(r'[a-zA-Z_]\w+$', file_name):
-        file_name = file_name + '.cfg'  # add .cfg suffix if not given
-    if not re.match(r'[a-zA-Z_]\w+\.cfg$', file_name):
-        raise ValueError(f'invalid file name given: {file_name}')
-    if (file_name == 'qteasy.cfg') and (not allow_default_name):
-        # TODO: 实现配置变量写入qteasy.cfg初始配置文件的功能
-        raise NotImplementedError(f'"qteasy.cfg" is not an allowed file name, functionality not '
-                                  f'implemented yet, please use another file name.')
-    return file_name
-
-
-def save_config(*, config=None, file_name=None, overwrite=True, initial_config=False) -> str:
-    """ 将config保存为一个文件
-    尚未实现的功能：如果initial_config为True，则将配置更新到初始化配置文件qteasy.cfg中()
-
-    Parameters
-    ----------
-    config: ConfigDict or dict, Default: None
-        一个config对象或者包含配配置变量的dict，如果为None，则保存qt.QT_CONFIG
-    file_name: str, Default: None
-        文件名，如果为None，文件名为"saved_config.cfg"
-    overwrite: bool, Default: True
-        默认True，覆盖重名文件，如果为False，当保存的文件已存在时，将报错
-    initial_config: bool, Default: False ** FUNCTIONALITY NOT IMPLEMENTED **
-        保配置变量到初始配置文件 qteasy.cfg 中，如果qteasy.cfg中已经存在部配置变量了，则覆盖相配置变量
-        TODO: 实现将配置变量写入qteasy.cfg初始配置文件的功能
-         由于目前使用pickle写入对象为二进制文件，而qteasy.cfg是文本文件，所以需要实现一个新的写入方式
-
-    Returns
-    -------
-    file_name: str
-        保存的文件名
-    """
-
-    from qteasy import logger_core
-    from qteasy import QT_ROOT_PATH
-    import os
-
-    if config is None:
-        config = QT_CONFIG
-    if not isinstance(config, (ConfigDict, dict)):
-        raise TypeError(f'config should be a ConfigDict or a dict, got {type(config)} instead.')
-
-    file_name = _check_config_file_name(file_name=file_name, allow_default_name=initial_config)
-
-    config_path = os.path.join(QT_ROOT_PATH, '../config/')
-
-    # write_binary_file
-    from .utilfuncs import write_binary_file
-    if overwrite:
-        open_method = 'wb'  # overwrite the file
-    else:
-        open_method = 'xb'  # raise if file already existed
-    try:
-        file_name = write_binary_file(
-                file_path=config_path,
-                file_name=file_name,
-                data=config,
-                mode=open_method,
-        )
-        logger_core.info(f'config file content written to: {file_name}')
-        return file_name
-    except Exception as e:
-        logger_core.warning(f'{e}, error during writing config to local file.')
-        return ""
-
-
-def load_config(*, config=None, file_name=None) -> dict:
-    """ 从文件file_name中读取相应的config参数，写入到config中，如果config为
-        None，则保存参数到QT_CONFIG中
-
-    Parameters
-    ----------
-    config: ConfigDict 对象
-        一个config对象，默认None，如果不为None且为一个ConfigDict对象，则将读取的配置参数写入到config中
-    file_name: str
-        文件名，默认None，如果为None，文件名为saved_config.cfg
-
-    Returns
-    -------
-    config: dict
-        读取的配置参数
-
-    Raises
-    ------
-    FileNotFoundError
-        如果给定文件不存在，则报错。如果没有给定文件名，则当config/saved_config.cfg不存在时，报错
-    """
-    from qteasy import logger_core
-    from qteasy import QT_ROOT_PATH
-
-    file_name = _check_config_file_name(file_name=file_name, allow_default_name=False)
-
-    config_path = os.path.join(QT_ROOT_PATH, '../config/')
-
-    # read_binary_file
-    from .utilfuncs import read_binary_file
-    try:
-        saved_config = read_binary_file(
-            file_path=config_path,
-            file_name=file_name,
-            mode='rb',
-        )
-        logger_core.info(f'read configuration file: {file_name}')
-    except FileNotFoundError as e:
-        msg = f'{e}\nConfiguration file {file_name} not found! nothing will be read.'
-        logger_core.warning(msg)
-        saved_config = {}
-
-    if config is not None:
-        if not isinstance(config, ConfigDict):
-            msg = f'config should be a ConfigDict, got {type(config)} instead.'
-            raise TypeError(msg)
-        configure(config=config,
-                  only_built_in_keys=False,
-                  **saved_config)
-
-    return saved_config
-
-
-def view_config_files(details=False) -> None:
-    """ 查看已经保存的配置文件，并显示其主要内容
-
-    Parameters
-    ----------
-    details: bool, Default: False
-        是否显示配置文件的详细内容
-
-    Returns
-    -------
-    None
-
-    Notes
-    -----
-    该函数只显示config文件夹下的配置文件，不显示qteasy.cfg中的配置
-    """
-
-    # TODO: add unittests
-    # 1. list all files in config folder
-    # 2. read each file and display its content if details is True
-    # 3. display only file names if details is False
-
-    from qteasy import QT_ROOT_PATH
-    import pickle
-
-    config_path = os.path.join(QT_ROOT_PATH, '../config/')
-
-    files = os.listdir(config_path)
-    files = [f for f in files if f.endswith('.cfg')]
-
-    if len(files) == 0:
-        print('No config files found in config folder.')
-
-    if details:
-        for file in files:
-            with open(os.path.join(config_path, file), 'rb') as f:
-                config = pickle.load(f)
-            print(f'File: {file}')
-            print(config)
-
-    else:
-        print('Config files found in config folder:')
-        for file in files:
-            print(file)
-
-
-def reset_config(config=None):
-    """ 重设config对象，将所有的参数都设置为默认值
-        如果config为None，则重设qt.QT_CONFIG
-
-    Parameters
-    ----------
-    config: ConfigDict
-        需要重设的config对象
-
-    Returns
-    -------
-    None
-
-    Notes
-    -----
-    等同于调用configure(config, reset=True)
-    """
-    from qteasy import logger_core
-    if config is None:
-        config = QT_CONFIG
-    if not isinstance(config, ConfigDict):
-        raise TypeError(f'config should be a ConfigDict, got {type(config)} instead.')
-    logger_core.info(f'{config} is now reset to default values.')
-    configure(config, reset=True)
 
 
 def live_trade_accounts() -> pd.DataFrame:
